+++ tests/Jest/services/ConfigService.spec.js
@@ -26,6 +26,22 @@ describe('TestRail Domain', () => {
         });
 
     });
+
+
+    test('Leading http:// is removed from TestRail domain', () => {
+        const config = new ConfigService({
+            'TESTRAIL_DOMAIN': 'http://my-domain',
+        });
+        expect(config.getDomain()).toBe('my-domain');
+    });
+
+    test('Leading https:// is removed from TestRail domain', () => {
+        const config = new ConfigService({
+            'TESTRAIL_DOMAIN': 'https://my-domain',
+        });
+        expect(config.getDomain()).toBe('my-domain');
+    });
+
 });
 
 describe('TestRail Username', () => {
@@ -316,6 +332,48 @@ describe('RunName', () => {
     });
 });
 
+describe('RunIncludeAll', () => {
+
+    describe('.env file', () => {
+
+        test('RunIncludeAll from .env ON', () => {
+            const config = new ConfigService({
+                'testrail': {
+                    'runIncludeAll': true,
+                },
+            });
+            expect(config.includeAllCasesDuringCreation()).toBe(true);
+        });
+
+        test('RunIncludeAll from .env OFF', () => {
+            const config = new ConfigService({
+                'testrail': {
+                    'runIncludeAll': false,
+                },
+            });
+            expect(config.includeAllCasesDuringCreation()).toBe(false);
+        });
+
+        test('RunIncludeAll from .env default is OFF', () => {
+            const config = new ConfigService({
+                'testrail': {},
+            });
+            expect(config.includeAllCasesDuringCreation()).toBe(false);
+        });
+
+    });
+
+    describe('ENV variable', () => {
+
+        test('RunIncludeAll from ENV variable', () => {
+            const config = new ConfigService({
+                'TESTRAIL_RUN_INCLUDE_ALL': true,
+            });
+            expect(config.includeAllCasesDuringCreation()).toBe(true);
+        });
+    });
+});
+
 describe('CloseRun', () => {
 
     describe('.env file', () => {
@@ -388,6 +446,48 @@ describe('Screenshots', () => {
     });
 });
 
+
+describe('Screenshots All', () => {
+
+    describe('.env file', () => {
+
+        test('ScreenshotsAll from .env ON', () => {
+            const config = new ConfigService({
+                'testrail': {
+                    'screenshotsAll': true,
+                },
+            });
+            expect(config.includeAllFailedScreenshots()).toBe(true);
+        });
+
+        test('ScreenshotsAll from .env OFF', () => {
+            const config = new ConfigService({
+                'testrail': {
+                    'screenshotsAll': false,
+                },
+            });
+            expect(config.includeAllFailedScreenshots()).toBe(false);
+        });
+
+        test('ScreenshotsAll from .env default is OFF', () => {
+            const config = new ConfigService({
+                'testrail': {},
+            });
+            expect(config.includeAllFailedScreenshots()).toBe(false);
+        });
+
+    });
+
+    describe('ENV variable', () => {
+
+        test('ScreenshotsAll from ENV variable', () => {
+            const config = new ConfigService({
+                'TESTRAIL_SCREENSHOTS_ALL': true,
+            });
+            expect(config.includeAllFailedScreenshots()).toBe(true);
+        });
+    });
+});
 
 describe('Invalid Configurations', () => {